requirements.txt
codecarbon
psutil
tree-sitter
tree-sitter-python
tree-sitter-javascript
requests
click
flask
pytest
pyyaml